LayoutTests/platform/ios-simulator/editing/pasteboard/line-feed-between-br-and-b-should-not-reorder-pasted-content-expected.txt

EDITING DELEGATE: shouldBeginEditingInDOMRange:range from 0 of DIV > BODY > HTML > #document to 3 of DIV > BODY > HTML > #document
EDITING DELEGATE: webViewDidBeginEditing:WebViewDidBeginEditingNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: shouldDeleteDOMRange:range from 1 of DIV > DIV > BODY > HTML > #document to 6 of #text > DIV > DIV > BODY > HTML > #document
EDITING DELEGATE: shouldChangeSelectedDOMRange:range from 0 of DIV > DIV > BODY > HTML > #document to 0 of DIV > DIV > BODY > HTML > #document toDOMRange:range from 0 of DIV > DIV > BODY > HTML > #document to 0 of DIV > DIV > BODY > HTML > #document affinity:NSSelectionAffinityDownstream stillSelecting:FALSE
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChange:WebViewDidChangeNotification
EDITING DELEGATE: shouldInsertNode:#document-fragment replacingDOMRange:range from 0 of DIV > DIV > BODY > HTML > #document to 0 of DIV > DIV > BODY > HTML > #document givenAction:WebViewInsertActionPasted
EDITING DELEGATE: shouldChangeSelectedDOMRange:range from 5 of #text > DIV > DIV > BODY > HTML > #document to 5 of #text > DIV > DIV > BODY > HTML > #document toDOMRange:range from 0 of DIV > DIV > BODY > HTML > #document to 5 of #text > DIV > DIV > BODY > HTML > #document affinity:NSSelectionAffinityDownstream stillSelecting:FALSE
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChange:WebViewDidChangeNotification
EDITING DELEGATE: shouldApplyStyle:<DOMCSSStyleDeclaration: 0x7fc61c950d10> toElementsInDOMRange:range from 0 of DIV > DIV > BODY > HTML > #document to 5 of #text > DIV > DIV > BODY > HTML > #document
EDITING DELEGATE: shouldChangeSelectedDOMRange:range from 5 of #text > SPAN > FONT > DIV > DIV > BODY > HTML > #document to 5 of #text > SPAN > FONT > DIV > DIV > BODY > HTML > #document toDOMRange:range from 0 of SPAN > FONT > DIV > DIV > BODY > HTML > #document to 5 of #text > SPAN > FONT > DIV > DIV > BODY > HTML > #document affinity:NSSelectionAffinityDownstream stillSelecting:FALSE
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChange:WebViewDidChangeNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidEndEditing:WebViewDidEndEditingNotification
Tests: 
Bug fix for <rdar://problem/4035648> REGRESSION (Mail): line feed in source HTML file causes bad copy/paste behavior 
In particular, the line feed that appears between br and b elements should not reorder the content in createMarkup.
Expected Results: 
Should see the text self-documenting itself correctly: 
bold
plain


Before cut:
| "
"
| <div>
|   class="editing"
|   id="test"
|   "
"
|   <#selection-anchor>
|   <br>
|   "
"
|   <b>
|     "bold"
|     <br>
|     "
"
|   "
plain<#selection-focus>
"
| "
"

After cut:
| "
"
| <div>
|   class="editing"
|   id="test"
|   <#selection-caret>
|   <br>
| "
"

After paste:
| "
"
| <div>
|   class="editing"
|   id="test"
|   <font>
|     size="3"
|     <span>
|       style="-webkit-text-size-adjust: auto; background-color: rgba(255, 255, 255, 0);"
|       <br>
|       <b>
|         "bold"
|         <br>
|       "plain<#selection-caret>"
|   <br>
| "
"
